+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**v0.0.2** (2011-06-16):
+
+  - Added dbrefHooks plugin
+  - Added dbref.fetch monkey patch
+  - Added basic and coercion examples
+
 **v0.0.1** (2011-06-06):
 
   - Initial version of the DBRef Plugin for Mongose 1.3.3